+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning].
 - Added `MockedHost` C++ class (in form of header-only `evmc::mocked_host` library)
   which can be used to emulate Host behavior when testing VM implementations.
   [#456](https://github.com/ethereum/evmc/pull/456)
+- In the Rust bindings added more type aliases (`MessageKind`, `MessageFlags`, `StatusCode`,
+  `StorageStatus`, `Revision`).
+  [#206](https://github.com/ethereum/evmc/pull/206)
 
 ### Changed